added more copy constructors and random fill.

+++ src/tensor.cu
@@ -9,6 +9,7 @@ float* tensor<DIMS>::get() const {
 
 template<int DIMS>
 void tensor<DIMS>::set_n_elems() {
+	
 	m_num_elements += m_size[0];
 };
 
@@ -20,7 +21,7 @@ tensor<DIMS>::tensor(const std::array<int, DIMS> t_size) : m_size(t_size) {
 
 template<int DIMS>
 __host__ __device__ 
-uint32_t tensor<DIMS>::get_n_elems() const {
+size_t tensor<DIMS>::get_n_elems() const {
 	return m_num_elements; 
 };
